+++ library-system/src/main/java/com/maids/library_system/borrowing_record/repositories/BorrowingRecordRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.maids.library_system.borrowing_record.entities.BorrowingRecord;
 
-public interface BorrowingRecordRepository extends CrudRepository<BorrowingRecord, Integer>{
+public interface BorrowingRecordRepository extends CrudRepository<BorrowingRecord, Long>{
 	
-	BorrowingRecord findByBookIdAndPatronId(int bookId,int patronId);
+	BorrowingRecord findByBookIdAndPatronId(long bookId,long patronId);
 }
