+++ library-system/src/main/java/com/maids/library_system/borrowing_record/services/BorrowingRecordService.java
@@ -2,6 +2,6 @@ package com.maids.library_system.borrowing_record.services;
 
 public interface BorrowingRecordService {
 	
-	int borrowABook (int bookId,int patronId);
-	int returnABook (int bookId,int patronId);
+	Long borrowABook (long bookId,long patronId);
+	Long returnABook (long bookId,long patronId);
 }
