+++ library-system/src/main/java/com/maids/library_system/borrowing_record/services/impl/BorrowingRecordServiceImpl.java
@@ -22,7 +22,7 @@ public class BorrowingRecordServiceImpl implements BorrowingRecordService {
 
 
     @Override
-    public int borrowABook(int bookId, int patronId) {
+    public Long borrowABook(long bookId, long patronId) {
         BorrowingRecord borrowingRecord = new BorrowingRecord();
         mapToBorrowingRecord(borrowingRecord, bookId, patronId);
         borrowingRecordRepository.save(borrowingRecord);
@@ -31,14 +31,14 @@ public class BorrowingRecordServiceImpl implements BorrowingRecordService {
 
     @Override
     @Transactional
-    public int returnABook(int bookId, int patronId) {
+    public Long returnABook(long bookId, long patronId) {
         BorrowingRecord borrowedBook = borrowingRecordRepository.findByBookIdAndPatronId(bookId, patronId);
         borrowedBook.setReturnDate(LocalDate.now());
         borrowingRecordRepository.save(borrowedBook);
         return borrowedBook.getId();
     }
 
-    private void mapToBorrowingRecord(BorrowingRecord borrowingRecord, int bookId, int patronId) {
+    private void mapToBorrowingRecord(BorrowingRecord borrowingRecord, long bookId, long patronId) {
         Book book = bookRepository.findById(bookId).orElseThrow(() -> new RuntimeException("Book not found"));
         Patron patron = patronRepository.findById(patronId).orElseThrow(() -> new RuntimeException("Patron not found"));
         borrowingRecord.setPatron(patron);
