+++ library-system/src/main/java/com/maids/library_system/patron/repositories/PatronRepository.java
@@ -1,9 +1,8 @@
 package com.maids.library_system.patron.repositories;
 
+import com.maids.library_system.patron.entities.Patron;
 import org.springframework.data.repository.CrudRepository;
 
-import com.maids.library_system.patron.entities.Patron;
-
-public interface PatronRepository extends CrudRepository<Patron, Integer>{
+public interface PatronRepository extends CrudRepository<Patron, Long>{
 
 }
